Add track index, volume OSC routes (send-only)
For our implementation, always represent tracks by their GUID, which is
encoded into the OSC address, similar to the canonical Reaper OSC
implementation.

This allows the consumer to tie its understanding of state directly to
the specific track, without having to keep track of that track's index
within the project.

This also allows the possibility of the arpad consumer having a
different track ordering than the reaper mixer order (not required, but
certainly useful).