refactor: localize osc route handling
Rather than handling the send and receive side independently, pull the
implementation of send and receive into an OscRoute trait and create
dataless structs to implement the trait for each route.

Now, the @osc-doc comments decorate the route and all the code that
deals with a particular route lives at the same place.

Potential future TO-DO: consider moving the routes into a module.