Add readonly OscRoute for track index
Track index is the index into the tracklist according to reaper's mixer.

In theory, this route could be made writeable to allow reordering tracks
but this could get tricky (what if you give an index greater than the
total number of tracks? what happens to the track currently at the
requested index?)