osc_route: fix track volume mapping
We want osc to communicate volume from 0.0 to 1.0 in a format that maps
the action of the faders (volume faders on control surfaces generally all
use the same dB-scaled positioning). Reaper provides functions to do
this mapping, but it's still a bit ugly.

For now, it is what it is. Will probably beautify the code later.