Implement polling for params not exposed by csurf
Reaper gives us hooks for much of what we would like to send over osc,
but we may wish to send other parameters that we do not have a hook for.
For these, we need to poll and check for changes.

The PollSource trait provides a way to define these and the PollManager
provides a way to perform the polling within csurf's run() method.